Some small refactors (#39)
* Some small refactors
* Added --tui
